Remove compute(int, int) since the perlin noise algorithm always expect doubles as input. Therefore, the simple implementation needs to explicit specify the frequency as input

+++ src/main/java/h13/noise/GradientNoise.java
@@ -29,6 +29,6 @@ public interface GradientNoise {
      * @param y The y-coordinate in the noise domain.
      * @return the computed gradient noise value at the specified noise domain coordinates
      */
-    double compute(int x, int y);
+    double compute(double x, double y);
 
 }
